init commit

+++ src/main/java/com/sparta/WeatherWear/repository/BoardImageRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BoardImageRepository extends JpaRepository<BoardImage, Integer> {
+public interface BoardImageRepository extends JpaRepository<BoardImage, Long> {
 }
