+++ src/main/java/com/sparta/WeatherWear/repository/BoardLikeRepository.java
@@ -5,5 +5,5 @@ import com.sparta.WeatherWear.entity.Board;
 import com.sparta.WeatherWear.entity.BoardLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardLikeRepository extends JpaRepository<BoardLike, Integer> {
+public interface BoardLikeRepository extends JpaRepository<BoardLike, Long> {
 }
