+++ src/main/java/com/sparta/WeatherWear/repository/ClosetRepository.java
@@ -4,5 +4,5 @@ package com.sparta.WeatherWear.repository;
 import com.sparta.WeatherWear.entity.Clothes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClosetRepository extends JpaRepository<Clothes, Integer> {
+public interface ClosetRepository extends JpaRepository<Clothes, Long> {
 }
